+++ CHANGELOG.md
@@ -26,10 +26,18 @@ N/A
 - Swift 3 support (currently not merged in master, available using branch `feature/swift3`). README is up to date, please report if you find any diffs
 - Some APIs have been updated to be more swifty, check out the README for more information
 
-**Extensions iOS removed:**   
+**Extensions removed:**
+
+- UIDevice: 
 
 ```
 class func deviceModel() -> String
+```
+
+- {UI/NS}Color: 
+
+```
+convenience init(hex: String, alpha: Float)
 ```
 
 ### Enhancements